requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==12.1.1
whitenoise==6.6.0
gunicorn==21.2.0
